Rename the compiled C extension 'zstd._zstd'.
This makes space for Python code to wrap around the low-level API.
In this patch, zstd/__init__.py just re-exports everything defined
by _zstd, but that will change in the next patches.
