+++ zstd/__init__.py
@@ -1,0 +1,2 @@
+from __future__ import absolute_import
+from ._zstd import *
